% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcraft.r
\name{setup_warcraft}
\alias{setup_warcraft}
\title{setup_warcraft}
\usage{
setup_warcraft(home)
}
\arguments{
\item{home}{Computer's home directory. You can find this by entering
\code{path.expand("~/")} into your console.}
}
\description{
setup_warcraft
}
\examples{
\dontrun{
## get home directory
home <- path.expand("~/")

## install audio and save key
warcraft::setup_warcraft(home)

## enter warcraft mode
library(warcraft)

## set for entire session
warcraft_mode(Inf)
}
}
