% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcraft.r
\docType{package}
\name{warcraft-package}
\alias{warcraft-package}
\title{warcraft: Turn on Warcraft Mode in your R environment}
\description{
Plays audio from the classic computer game while you work: a "job's
done" when a task took real effort, an annoyed orc when it failed, and
random peon chatter the rest of the time. Requires \code{afplay}
(macOS) or one of \code{play}, \code{paplay}, or \code{aplay} (linux).
}
\section{Options}{

\describe{
  \item{\code{warcraft.mute}}{Set to \code{TRUE} to silence every clip.}
  \item{\code{warcraft.volume}}{Playback volume, where \code{1} is normal.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/warcraft}
  \item Report bugs at \url{https://github.com/mkearney/warcraft/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{mkearney@ku.edu}

Authors:
\itemize{
  \item Michael W. Kearney \email{mkearney@ku.edu}
}

}
