% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcraft.r
\docType{package}
\name{warcraft}
\alias{warcraft}
\alias{.onAttach}
\alias{warcraft-package}
\title{warcraft}
\usage{
.onAttach(libname, pkgname)
}
\description{
Set your R environment to warcraft mode
}
\examples{
\dontrun{
## for setup instructions see README which you can find
## https://github.com/mkearney/warcraft/blob/master/README.md
## to enter warcraft mode
library(warcraft)

## to re-enter or to prevent expiration in a session
warcraft_mode(Inf)
}

}
