% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcraft.r
\docType{package}
\name{warcraft-package}
\alias{warcraft}
\alias{warcraft-package}
\title{warcraft: Turn on Warcraft Mode in your R environment}
\description{
Playing audio files from the classic computer game during
random top-level R tasks. Sorry, only works for Macs/linux
based operating systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/warcraft}
  \item Report bugs at \url{https://github.com/mkearney/warcraft/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{mkearney@ku.edu}

}
