% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcraft.r
\name{warcraft_mode}
\alias{warcraft_mode}
\title{warcraft mode}
\usage{
warcraft_mode(n = 500)
}
\arguments{
\item{n}{Number of max times to call warcraft audio before
warcraft mode expires. Defaults to 500. Plays roughly 1
out of 10 top level tasks.}
}
\description{
warcraft mode
}
