% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcraft.r
\name{warcraft_mode}
\alias{warcraft_mode}
\title{jobsdone_wrappper}
\usage{
warcraft_mode(n = 100)
}
\arguments{
\item{n}{Number of max times to call warcraft audio before
warcraft mode expires. Defaults to 100. Plays roughly 1
out of 10 top level tasks.}
}
\description{
jobsdone_wrappper
}
