% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{warcraft_mode}
\alias{warcraft_mode}
\alias{warcraft}
\title{Enter Warcraft Mode}
\usage{
warcraft_mode(n = 500, p = 0.1)

warcraft(n = 500, p = 0.1)
}
\arguments{
\item{n}{Maximum number of times to call warcraft audio before
warcraft mode expires. Defaults to 500. To disable warcraft mode,
set this value to FALSE.}

\item{p}{Desired proportion of top-level tasks that should trigger
the playing of audio files. Defaults to play audio during roughly
1 out of 10 top level tasks.}
}
\value{
Invisibly enters warcraft_mode (getOption("warcraft_mode"))
}
\description{
Enter Warcraft Mode
}
\examples{
\dontrun{
## load warcraft
library(warcraft)

## enter warcraft mode
warcraft_mode()

## disable warcraft mode
warcraft_mode(FALSE)

## play lots of warcraft sounds for entire R session
warcraft_mode(Inf, 1.0)

}
}
